refactor: separate postalb copyright notice into its own component
