+++ src/components/TeacherComponents/CopyrightLicence/CopyrightLicence.test.tsx
@@ -1,0 +1,75 @@
+import { screen } from "@testing-library/dom";
+
+import CopyrightLicence from "./CopyrightLicence";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
+describe("CopyrightLicence", () => {
+  it("renders post-ALB copyright notice", () => {
+    renderWithTheme(
+      <CopyrightLicence
+        openLinksExternally={false}
+        copyrightYear="2024-01-01T00:00:00Z"
+      />,
+    );
+
+    const preAlbCopyright = screen.getByText(
+      `This content is © Oak National Academy Limited (2024), licensed on`,
+      { exact: false },
+    );
+    expect(preAlbCopyright).toBeInTheDocument();
+  });
+  it("opens links in a new tab when openLinksExternally is true", async () => {
+    renderWithTheme(
+      <CopyrightLicence
+        openLinksExternally={true}
+        copyrightYear="2024-01-01T00:00:00Z"
+      />,
+    );
+
+    const links = screen.getAllByRole("link");
+    const firstLink = links[0];
+    if (!firstLink) {
+      throw new Error("Failed to find any links in CopyrightNotice component");
+    }
+    expect(firstLink).toHaveAttribute("target", "_blank");
+  });
+  it("opens links in the same tab when openLinksExternally is false", async () => {
+    renderWithTheme(
+      <CopyrightLicence
+        openLinksExternally={false}
+        copyrightYear="2024-01-01T00:00:00Z"
+      />,
+    );
+
+    const links = screen.getAllByRole("link");
+    const firstLink = links[0];
+    if (!firstLink) {
+      throw new Error("Failed to find any links in CopyrightNotice component");
+    }
+    expect(firstLink).not.toHaveAttribute("target", "_blank");
+  });
+  it("links that will open in a new tab should be tagged with an icon", () => {
+    renderWithTheme(
+      <CopyrightLicence
+        openLinksExternally={true}
+        copyrightYear="2024-01-01T00:00:00Z"
+      />,
+    );
+    const links = screen.getAllByRole("link");
+    const externalLinkIcons = screen.getAllByTestId("external-link-icon");
+    expect(links).toHaveLength(externalLinkIcons.length);
+  });
+  it("links that do not open in a new tab should not be tagged with an icon", () => {
+    renderWithTheme(
+      <CopyrightLicence
+        openLinksExternally={false}
+        copyrightYear="2024-01-01T00:00:00Z"
+      />,
+    );
+    const links = screen.getAllByRole("link");
+    expect(links).toHaveLength(2);
+    const externalLinkIcons = screen.queryAllByTestId("external-link-icon");
+    expect(externalLinkIcons).toHaveLength(0);
+  });
+});
